+++ src/main/java/com/fitTracker/fitTracker/Service/impl/EnderecoServiceImpl.java
@@ -3,6 +3,7 @@ package com.fitTracker.fitTracker.Service.impl;
 import com.fitTracker.fitTracker.Models.Endereco;
 import com.fitTracker.fitTracker.Repositories.EnderecoRepository;
 import com.fitTracker.fitTracker.Service.EnderecoService;
+import com.fitTracker.fitTracker.Util.ElementoNaoEncontradoException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -26,11 +27,23 @@ public class EnderecoServiceImpl implements EnderecoService {
 
     @Override
     public Optional<Endereco> findById(Long id) {
-        return enderecoRepository.findById(id);
+
+        Optional<Endereco> endereco = enderecoRepository.findById(id);
+
+        if(endereco.isEmpty()) {
+            throw new ElementoNaoEncontradoException("Não foi encontrado nenhum Endereco com esse id");
+        }
+
+        return endereco;
     }
 
     @Override
     public void deleteById(Long id) {
+
+        if(enderecoRepository.findById(id).isEmpty()) {
+            throw new ElementoNaoEncontradoException("Não foi encontrado nenhum Endereco com esse id");
+        }
+
         enderecoRepository.deleteById(id);
     }
 }
